added unit tests project, removed String::add() methods for types other than text, added String::insert/remove/clear(), replaced TextView's and TextField's custom layouts with IDWriteTextLayout, renamed ScopedPointer to SharedPointer

+++ Engine/src/application/Timer.cpp
@@ -14,10 +14,10 @@ namespace Ghurund {
         QueryPerformanceCounter(&currentFrame);
         ticks = currentFrame.QuadPart - startFrame.QuadPart;
         time = (double)ticks / frequency.QuadPart;
-        timeMs = ticks / (frequency.QuadPart / 1000.0);
+        timeMs = (uint32_t)(ticks / (frequency.QuadPart / 1000.0));
         frameTicks = currentFrame.QuadPart - prevFrame.QuadPart;
         frameTime = (float)frameTicks / frequency.QuadPart;
-        frameTimeMs = ticks / (frequency.QuadPart / 1000.0);
+        frameTimeMs = (uint32_t)(ticks / (frequency.QuadPart / 1000.0));
     }
 
     void Timer::setPaused(bool p) {
