+++ Engine/src/net/Socket.cpp
@@ -62,8 +62,8 @@ namespace Ghurund {
 
     Status Socket::send(const void* data, size_t size, unsigned int flags) const {
         if (addressStruct6 != nullptr)
-            return ::sendto(id, (const char*)data, size, flags, (sockaddr*)addressStruct6, sizeof(sockaddr_in6)) == SOCKET_ERROR ? Status::SOCKET : Status::OK;
-        return ::sendto(id, (const char*)data, size, flags, (sockaddr*)addressStruct, sizeof(sockaddr_in)) == SOCKET_ERROR ? Status::SOCKET : Status::OK;
+            return ::sendto(id, (const char*)data, (int)size, flags, (sockaddr*)addressStruct6, sizeof(sockaddr_in6)) == SOCKET_ERROR ? Status::SOCKET : Status::OK;
+        return ::sendto(id, (const char*)data, (int)size, flags, (sockaddr*)addressStruct, sizeof(sockaddr_in)) == SOCKET_ERROR ? Status::SOCKET : Status::OK;
     }
 
     Status Socket::receive(void** data, size_t * size) {
